+++ mithril-stm/src/aggregate_signature/mod.rs
@@ -259,6 +259,60 @@ mod tests {
                     unreachable!()
             }
         }
+
+        #[test]
+        /// Test that batch verification of certificates works
+        fn batch_verify(nparties in 2_usize..30,
+                              m in 10_u64..20,
+                              k in 1_u64..4,
+                              seed in any::<[u8;32]>(),
+                              batch_size in 2..10,
+        ) {
+            let mut rng = ChaCha20Rng::from_seed(seed);
+            let mut aggr_avks = Vec::new();
+            let mut aggr_stms = Vec::new();
+            let mut batch_msgs = Vec::new();
+            let mut batch_params = Vec::new();
+            for _ in 0..batch_size {
+                let mut msg = [0u8; 32];
+                rng.fill_bytes(&mut msg);
+                let params = StmParameters { m, k, phi_f: 0.95 };
+                let ps = setup_equal_parties(params, nparties);
+                let clerk = StmClerk::from_signer(&ps[0]);
+
+                let all_ps: Vec<usize> = (0..nparties).collect();
+                let sigs = find_signatures(&msg, &ps, &all_ps);
+                let msig = clerk.aggregate(&sigs, &msg);
+
+                match msig {
+                    Ok(aggr) => {
+                        aggr_avks.push(clerk.compute_avk());
+                        aggr_stms.push(aggr);
+                        batch_msgs.push(msg.to_vec());
+                        batch_params.push(params);
+                    }
+                    Err(AggregationError::NotEnoughSignatures(_n, _k)) => {
+                        assert!(sigs.len() < params.k as usize)
+                    }
+                    Err(AggregationError::UsizeConversionInvalid) => unreachable!(),
+                }
+            }
+
+            assert!(StmAggrSig::batch_verify(&aggr_stms, &batch_msgs, &aggr_avks, &batch_params).is_ok());
+
+            let mut msg = [0u8; 32];
+            rng.fill_bytes(&mut msg);
+            let params = StmParameters { m, k, phi_f: 0.8 };
+            let ps = setup_equal_parties(params, nparties);
+            let clerk = StmClerk::from_signer(&ps[0]);
+
+            let all_ps: Vec<usize> = (0..nparties).collect();
+            let sigs = find_signatures(&msg, &ps, &all_ps);
+            let fake_msig = clerk.aggregate(&sigs, &msg);
+
+            aggr_stms[0] = fake_msig.unwrap();
+            assert!(StmAggrSig::batch_verify(&aggr_stms, &batch_msgs, &aggr_avks, &batch_params).is_err());
+        }
     }
 
     proptest! {